docs/logic-tool/requirements.txt
requests
python-dotenv
fastapi
uvicorn
geopy